Support compiling k2 on macOS/Linux/Windows without installing CUDA toolkit. (#748)
* Support compiling k2 on macOS/Linux/Windows without installing CUDA toolkit.

* Update version to 0.3.5

+++ k2/python/tests/fsa_test.py
@@ -27,7 +27,7 @@ class TestFsa(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.devices = [torch.device('cpu')]
-        if torch.cuda.is_available():
+        if torch.cuda.is_available() and k2.with_cuda:
             cls.devices.append(torch.device('cuda', 0))
 
     def test_acceptor_from_tensor(self):
@@ -682,8 +682,10 @@ class TestFsa(unittest.TestCase):
         fsa.symbols = symbols
         fsa.aux_symbols = aux_symbols
 
-        fsa.draw(filename='foo.png')
-        os.remove('foo.png')
+        import shutil
+        if shutil.which('dot') is not None:
+            fsa.draw(filename='foo.png')
+            os.remove('foo.png')
 
     def test_to(self):
         s = '''
@@ -693,7 +695,7 @@ class TestFsa(unittest.TestCase):
         fsa = k2.Fsa.from_str(s)
         assert fsa.is_cpu()
 
-        if torch.cuda.is_available():
+        if torch.cuda.is_available() and k2.with_cuda:
             fsa = fsa.to('cuda:0')
             assert fsa.is_cuda()
